Add files via upload
The used dataset contains transaction details of fraudulent and normal transactions. Most attributes have undergone PCA analysis in order to protect confidential details such as credit card number, cvv, expiry date and so on. 

The aim of this project is to conduct preliminary analysis on the given 'creditcard.csv' dataset to make initial inferences and develop a model to analyse the data. The dataset is divided into fraudulent and normal transactions which the model should successfully identify when run on the test set. Two techniques are to be tested and then compared accordingly with the help of performanec metrics.